Revert "ai: add cfg_brains with CfgBrains suppression tweaks"
Further testing has shown that it's way to easy to suppress
AI soldiers with stray bullets in busy combat scenarios, making
them essentially unable to hit even at close range, and there's
no working tunable to reduce the radius of nearby bullets passing
by to tweak this.

AI still gets suppressed and gets reduced accuracy as a result
of that, even with vanilla defaults (just not as much), and it
wears off much quicker.

This reverts commit 611da91d2cca6672327b10a1e4d06bd20626f342.