map_trackers: keep updating on briefing screen
The uiSleep ensures that we'll continue refreshing markers
as new elements join in, potentially leaving back to lobby,
and re-joining again.

The null check ensures we don't process unused groups - the engine
apparently pre-creates all groups and units from the lobby, even
if left empty, with objNull units. This resulted in ghost group
markers on [0,0,0].

Signed-off-by: freghar <freghar@dummy.tld>