infopanel_control: remove, was mostly unused
This was a way to prevent players from ever opening up / seeing
a GPS screen or a mine detector screen, regardless whether they
had a GPS item or whether a vehicle they were in provided one
(such as vanilla offroads).

The logic this code used, however, wasn't ideal and often didn't
work due to different names used for left/right panels across
vehicles (ie. VehicleSystemsDisplayManagerComponentLeft).

So if you really need to disable GPS for some vehicle that has
it by default, use enableInfoPanelComponent manually on that
vehicle (or one of its turrets), with a name extracted from
infoPanels or the vehicle config.

Signed-off-by: freghar <freghar@dummy.tld>