remoteExec 3den object attribute if run on non-server
This works around a fairly rare engine bug, which causes the 3den
object attribute expression= to execute on an existing in-game client
(rather than the server) on JIP of a new client.

This might have been some hidden load balancing feature (creating
JIPed units on other clients, not the server), but it's undocumented
and annoying.

Signed-off-by: freghar <freghar@dummy.tld>