ee/acre_channels: remove buggy diary note
The issue is that groupId isn't set or propagated at the time
of diary entry addition, resulting in ie. "Alpha-1-1" being used
instead.

The "proper" way would be to somehow wait for setGroupIdGlobal
to propagate, but since that still leaves a lot of JIP queue
entries, just remove the functionality. It wasn't worth it.

Signed-off-by: freghar <freghar@dummy.tld>