add addonRootClass to anything adding CfgVehicles/Weapons
This is not how it's supposed to be used, and it breaks the idea behind
requiredAddons[] in missions and custom compositions, but it gives me
flexibility to move/rename addons (while retaining classnames) without
missions breaking.

And if I rename a classname, at least the mission can be repaired
without having to unRap the mission.sqm, because the editor will still
load it (despite a warning popup).

Signed-off-by: freghar <freghar@dummy.tld>