NOTES
addonRootClass works by inserting the current CfgPatches class into the
referenced one, unfortunately including requiredAddons[].

This means that we can't ie. modify rhs_blabla while pointing at it via
addonRootClass, because it would put the dependency inside rhs_blabla
and result in:

class CfgPatches {
  class rhs_blabla {
    ...
    requiredAddons[] =  {
      ...
      "rhs_blabla"
    };
    ...
  };
};

This creates a circular dep, causing the game to crash on start.

The correct approach would be to find a CfgPatches class, which is
part of standard base Arma3 and upon which NO MOD DEPENDS, and use it
as addonRootClass.

This will put our requiredAddons[] inside that base Arma3 class, ie.
making it depend on "rhs_blabla", but as long as rhs_blabla doesn't
depend on this base Arma3 class, no circular dep is created.

The mission will then depend on the base Arma3 class, so we don't
get rid of a mission dependency completely, but at least the mission
will load on both vanilla and modded game.

The only real problem is finding this exotic base Arma3 class upon
which noone ever depends. "Core" seems to be a good candidate.

The ultimate question is then - is this worth it? Is it worth the
confusing error messages when ie. RHS isn't loaded? ("Core depends
on rhs_blabla which is not loaded") instead of our addon name.

I don't think so (for now).