replace CBA PFH with Arma v2.03 EachFrame with arg
Removing Ambient Anim EH is not easily possible as multiple modules
add units to the one loop, which starts as empty (and thus an
removeEventHandler would immediately remove it).

Since there's no central place from which we could count all modules
and then increment some global var and thus check if all modules
are done waiting for combat, .. we cannot easily check that all modules
are, in fact, done.

So just leave the loop running.

Signed-off-by: freghar <freghar@dummy.tld>